test: move the repo-hygiene checks out of the root package
CI lint failed with SA5011 in repo_format_test.go, on code byte-identical
to main, where the same golangci-lint version passes.

The trigger was these two tests. They were written in package cloudstic
and import golang.org/x/tools/go/packages, which grew that package's
type-check graph from 370 to 423 packages — enough for staticcheck to
lose the transitive fact that testing.T.Fatal never returns, so a
guarded pointer read three files away was reported as a possible nil
dereference.

They never needed to live there: both work by loading packages via
import path, not by importing the package under test. Moving them to
internal/apicheck takes x/tools back out of the root package's test
binary.

goreleaser_test.go read .goreleaser.yml relative to the working
directory, which the move changes, so it now walks up to the directory
holding go.mod.

Both guards were re-verified after the move by breaking them on purpose:
a typo in a -X symbol name and a removed FileMeta alias each produce the
expected failure.
